moved interactive plots to separate file
